@@ -1,27 +1,26 @@
 source :rubygems
 
-gem 'rails', '3.0.10'
-gem 'thin', '~>1.3.1'
-gem 'will_paginate', '~> 2.3.11'
-gem 'mysql2', '~> 0.2.0'
 
+gem 'rails', '~> 3.2.8'
+gem 'thin', '~>1.3.1'
+gem 'will_paginate', '~> 3.0'
+gem 'mysql2', '~> 0.3.0'
 gem 'newrelic_rpm','~>3.0.1'
 gem 'json', '~> 1.6.1'
 gem 'nokogiri', '~> 1.4.4'
-gem "activemerchant", '~> 1.15.0', :require => "active_merchant"
 gem 'uakari', '~> 0.2.0'
 gem "pusher", "~> 0.9.2"
 gem "aws-s3", "~> 0.6.2", :require => "aws/s3"
 gem "paperclip", "2.3.16"
 gem 'daemons', '~> 1.0.10'
-gem 'twilio', '~> 3.0.1'
+gem 'twilio', '~> 3.1.0'
 gem 'settingslogic', '~> 2.0.6'
 gem "twilio-ruby", '~> 3.5.1'
 gem 'dynamic_form', '~> 1.1.4'
 gem 'in_place_editing', '~> 1.1.1'
 gem "nested_form", '~>0.1.1'
 gem 'jquery-rails', '>= 1.0.12'
-gem 'delayed_job', '~>2.1.4'
+gem 'delayed_job'
 gem 'heroku','~>2.17.0'
 gem "activerecord-import", ">= 0.2.0"
 gem "airbrake",'~>3.0.9'
@@ -31,7 +30,6 @@ gem "call_center", "~> 0.0.9"
 gem "em-http-request", "~> 1.0.1"
 gem "ruby-graphviz", "0.9.0"
 gem "em-synchrony", "~> 1.0.1"
-gem "rack_timer", "0.0.4"
 gem "resque", "~> 1.19.0"
 gem "resque-scheduler", "~> 1.9.9", :require => 'resque_scheduler'
 gem 'resque-lock', "~> 1.0.0"
@@ -40,21 +38,11 @@ gem "redis", "~>2.2.2"
 gem "hiredis", "~>0.4.5"
 gem "formtastic", "~>2.2.1"
 gem "cocoon", "~>1.0.22"
-gem "sidekiq", "~>1.2.1"
-gem "slim", "~>1.2.2"
-gem "sinatra", "~>1.2.8"
-gem "sprockets", "~>2.4.5"
-
-
-
-
-
 
 group :development, :test do
   gem 'rspec'
-  gem 'ruby-debug19'
   gem 'rspec-rails', '~> 2.6.1'
-
+  gem 'debugger'
 end
 
 group :development do
@@ -63,15 +51,17 @@ group :development do
   gem 'guard-rspec'
   gem 'guard-spork'
   gem 'guard-rails'
-  gem 'capistrano', '2.9.0'
+  gem 'rb-fsevent'
   gem 'heroku_san'
+  gem 'showoff-io'
+  gem 'foreman'
 end
-
-
 
 group :test do
   gem 'factory_girl', '~> 1.3.3'
-  gem 'shoulda', '~> 2.11.3'
+  gem 'shoulda'
   gem 'simplecov'
-  gem 'spork', '~> 0.9.0.rc9'
+  gem 'spork-rails'
+  gem "database_cleaner"
+  gem "json_spec"
 end
