@@ -1,44 +1,114 @@
 module Client
-  class ScriptsController < ::ScriptsController
-    skip_before_filter :load_script, :apply_changes, :questions_answered
+  class ScriptsController < ClientController
+    before_filter :load_and_verify_script, :except => [:index, :new, :create, :deleted]
+    before_filter :load_voter_fields, :only => [ :show, :edit]
 
     layout 'client'
 
-    def deleted
-      render 'scripts/deleted'
-    end
+    respond_to :html, :json
 
     def index
-      @scripts = @user.scripts.manual.active.paginate(:page => params[:page])
+      @scripts = account.scripts.manual.active.paginate(:page => params[:page])
+      respond_with @scripts      
     end
-
+    
+    def show
+      respond_with @script do |format|
+        format.html {redirect_to edit_client_script_path(@script)}
+      end      
+    end
+    
+    def edit
+      respond_with @script
+    end
+    
     def new
-      @script = Script.new(:robo => false, questions: [Question.new(possible_responses: [PossibleResponse.new])], script_texts: [ScriptText.new])
-      @voter_fields = VoterList::VOTER_DATA_COLUMNS.values
-      @voter_fields.concat(@user.account.custom_voter_fields.collect{ |field| field.name})
-      @voter_field_values=[]
+      new_script
+      load_voter_fields
+      @script.script_texts.new(script_order: 1)
+      @question = @script.questions.new(script_order: 2)
+      @question.possible_responses.new(possible_response_order: 1)
+      respond_with @script
     end
 
     def create
-      params[:script][:voter_fields] =  params[:voter_field] ? params[:voter_field].to_json : nil
-      @script = Script.new(params[:script])
-      @voter_fields = Voter.upload_fields
-      @voter_field_values = params[:voter_field] || []
-      if @script.save
-        @user.account.scripts << @script
-        flash_message(:notice, "Script saved")
-        redirect_to :action=>"index"
-      else
-        render :action=>"new"
-      end
+      new_script
+      save_script
+      load_voter_fields
+      respond_with @script, location: client_scripts_path
     end
 
-    def show
-      @script = @user.account.scripts.find(params[:id])
-      @script.questions << [Question.new(possible_responses: [PossibleResponse.new])]  if @script.questions.empty?
+
+    def edit
+      respond_with @script
+    end
+
+    def update
+      save_script
+      respond_with @script,  location: client_scripts_path do |format|         
+        format.json { render :json => {message: "Script updated" }, :status => :ok } if @script.errors.empty?
+      end            
+      
+    end
+
+    def destroy
+      @script.active = false
+      @script.save ?  flash_message(:notice, "Script deleted") : flash_message(:error, @script.errors.full_messages.join)            
+      respond_with @script,  location: client_scripts_path do |format|         
+        format.json { render :json => {message: "Script deleted" }, :status => :ok } if @script.errors.empty?
+      end      
+    end
+
+    def questions_answered      
+      render :json => { :data => Question.question_count_script(@script.id) }
+    end
+
+    def possible_responses_answered
+      render :json => { :data => PossibleResponse.possible_response_count(params[:question_ids]) }
+    end
+
+    def deleted
+      @scripts = Script.deleted.for_account(account).paginate(:page => params[:page], :order => 'id desc')
+      respond_with @scripts do |format|
+        format.html{render 'scripts/deleted'}
+        format.json {render :json => @scripts.to_json}
+      end      
+    end
+    
+    def restore
+      @script.active = true
+      save_script
+      respond_with @script,  location: client_scripts_path do |format|         
+        format.json { render :json => {message: "Script restored" }, :status => :ok } if @script.errors.empty?
+      end      
+    end
+    
+
+
+
+    private
+    
+    def load_and_verify_script
+      begin
+        @script = Script.find(params[:id] || params[:script_id])
+      rescue ActiveRecord::RecordNotFound => e
+        render :json=> {"message"=>"Resource not found"}, :status => :not_found
+        return
+      end
+      if @script.account != account
+        render :json => {message: 'Cannot access script.'}, :status => :unauthorized
+        return
+      end
+    end
+    
+
+    def new_script
+      @script = account.scripts.new(robo: false)
+    end
+
+    def load_voter_fields
       @voter_fields = VoterList::VOTER_DATA_COLUMNS.values
       @voter_fields.concat(@user.account.custom_voter_fields.collect{ |field| field.name})
-      @answered_questions = Question.question_count_script(@script.id)
       if @script.voter_fields!=nil
         begin
           @voter_field_values = JSON.parse(@script.voter_fields)
@@ -49,82 +119,12 @@ module Client
         @voter_field_values=[]
       end
     end
-    
 
-    def update
-      @script = @user.account.scripts.find(params[:id])
-      params[:script][:voter_fields] =  params[:voter_field] ? params[:voter_field].to_json : nil
-      @voter_fields = VoterList::VOTER_DATA_COLUMNS.values
-      @voter_fields.concat(@user.account.custom_voter_fields.collect{ |field| field.name})
-      @voter_field_values = (JSON.parse(@script.voter_fields) if @script.voter_fields) || []
-      @answered_questions = Question.question_count_script(@script.id)
-      begin
-        params[:save_as] ? save_as : @script = account.scripts.find_by_id(params[:id])
-        if params[:save_as]
-          redirect_to client_script_path(@script)          
-        elsif !params[:save_as] && @script.update_attributes!(params[:script])
-          flash_message(:notice, "Script updated")
-          redirect_to :action=>"index"
-        else
-          render :show
-        end
-      rescue Exception => e
-        flash_message(:notice, "Script not saved. Error:" + e.message)
-        render :show
+    def save_script
+      unless params[:script].nil?
+        params[:script][:voter_fields] =  params[:voter_field] ? params[:voter_field].to_json : nil
       end
+      flash_message(:notice, "Script saved") if @script.update_attributes(params[:script])
     end
-
-    def destroy
-      @script = @user.account.scripts.manual.find(params[:id])
-      unless @script.nil?
-        campaign = @user.account.campaigns.active.find_by_script_id(@script.id)
-        if campaign.nil?
-          @script.update_attributes(:active => false)
-          flash_message(:notice, "Script deleted")
-        else
-          flash_message(:notice, I18n.t(:script_cannot_be_deleted))
-        end
-      end
-      redirect_to :action => "index"
-    end
-    
-    def questions_answered
-      render :json => { :data => Question.question_count_script(params[:id]) }
-    end
-    
-    def possible_responses_answered
-      render :json => { :data => PossibleResponse.possible_response_count(params[:question_ids]) }
-    end
-    
-    def load_deleted
-      self.instance_variable_set("@#{type_name.pluralize}", Script.deleted.for_account(@user.account).paginate(:page => params[:page], :order => 'id desc'))
-    end    
-    
-
-    private
-
-    def save_as
-      @script = Script.new(:name => "", :active => true, :account => @user.account, :script => params[:script][:script], :voter_fields => params[:script][:voter_fields])
-      @script.save(:validate => false)
-      params[:script][:questions_attributes].each_value.each do |q|
-        if q[:_destroy] == "false"
-          question = @script.questions.new(:text => q[:text])
-          question.save!
-          q[:possible_responses_attributes].each_value.each do |ps|
-            if ps[:_destroy] == "false"
-              possible_response = question.possible_responses.new(:value => ps[:value], :keypad => ps[:keypad], :retry => ps[:retry])
-              possible_response.save!
-            end
-          end
-        end
-      end
-      params[:script][:notes_attributes].try(:each_value).try(:each) do |n|
-        if n[:_destroy] == "false"
-          note = @script.notes.new(:note => n[:note])
-          note.save!
-        end
-      end
-    end
-
   end
 end
