@@ -2,14 +2,32 @@ require Rails.root.join("lib/twilio_lib")
 
 class ClientController < ApplicationController
   protect_from_forgery :except => [:billing_updated, :billing_success]
+  before_filter :authenticate_api
   before_filter :check_login, :except => [:login, :user_add, :forgot]
   before_filter :check_paid
 
   layout "client"
   in_place_edit_for :campaign, :name
-  
+
+
+  def authenticate_api
+    unless params[:api_key].blank?
+      @account = Account.find_by_api_key(params[:api_key])
+      return if @account.nil?
+      @user = @account.users.first
+      session[:user] = @user.id
+    end
+  end
+
+
   def check_login
-    redirect_to_login and return if session[:user].blank?
+    if session[:user].blank?
+      respond_to do |format|
+        format.json { render :json => {status: 'error', code: '401' , message: 'Unauthorized'}, :status => :unauthorized }
+        format.html { redirect_to login_path }
+      end
+      return
+    end
     begin
       @user = User.find(session[:user])
       @account = @user.account
@@ -44,8 +62,8 @@ class ClientController < ApplicationController
       end
     end
   end
-  
-  def caller_password    
+
+  def caller_password
     password = params[:caller_password]
     if password.blank? || password.length < 5
       flash_message(:error, "The Account caller password can't be less than 5 characters.")
@@ -56,17 +74,27 @@ class ClientController < ApplicationController
     redirect_to :back
   end
 
+  def generate_api_key
+    @account.enable_api!
+    redirect_to :back
+  end
+
+  def login_from_api_key
+    self.current_user = Account.find_by_api_key(params[:api_key]) unless params[:api_key].empty?
+  end
+
+
   def user_add
     @breadcrumb = "My Account"
     @title = "My Account"
 
     if session[:user].blank?
-      @user = User.new(:account => Account.new(:domain => request.domain), role: User::Role::ADMINISTRATOR)
+      @user = User.new(:account => Account.new(:domain_name => request.domain), role: User::Role::ADMINISTRATOR)
     else
       @user = User.find(session[:user])
       @account = @user.account
     end
-    
+
     if request.post?
       @user.attributes =  params[:user]
       if params[:fullname]!=nil
@@ -80,9 +108,10 @@ class ClientController < ApplicationController
         flash_now(:error, "Current password incorrect")
         return
       else
+        puts @user.inspect
         @user.save
       end
-      
+
       if @user.valid?
         @user.send_welcome_email
         @user.create_default_campaign
@@ -91,7 +120,7 @@ class ClientController < ApplicationController
         if session[:user].blank?
           message = "Your account has been created."
           session[:user]=@user.id
-          flash_message(:notice, message)          
+          flash_message(:notice, message)
           flash_message(:kissmetrics, "Signed Up")
           redirect_to :action=>"welcome"
           return
@@ -130,7 +159,7 @@ class ClientController < ApplicationController
 
   def login
     if session[:user]
-      redirect_to :action => "user_add" 
+      redirect_to :action => "user_add"
       return
     end
 
@@ -205,7 +234,7 @@ class ClientController < ApplicationController
      @recharge_options.tap{
       10.step(500,10).each do |n|
         @recharge_options << ["$#{n}",n.to_f]
-       end 
+       end
      }
    end
 
@@ -228,7 +257,7 @@ class ClientController < ApplicationController
      @user.account.sync_subscription
      redirect_to :action=>"billing"
    end
-   
+
    def billing_form
      @account_code=@user.account.recurly_account_code
      @billing_info = Recurly::Account.find(@account_code).billing_info
@@ -280,546 +309,6 @@ class ClientController < ApplicationController
    def new_subscription
      render :layout=>"recurly"
    end
-
-   def billing_old
-     @breadcrumb="Billing"
-     @billing_account = @user.billing_account || @user.account.new_billing_account
-     @oldcc = @billing_account.cc
-     if @billing_account.last4.blank?
-       @tempcc = ""
-       @billing_account.cc = ""
-     else
-       @tempcc = "xxxx xxxx xxxx #{@billing_account.last4}"
-       @billing_account.cc = "xxxx xxxx xxxx #{@billing_account.last4}"
-     end
-
-     if request.post?
-       @billing_account.account_id = account.id
-       @billing_account.attributes = params[:billing_account]
-
-       if @billing_account.cc==@tempcc
-         @billing_account.cc = @oldcc
-       else
-         if @billing_account.cc.length > 4
-           @billing_account.last4 = @billing_account.cc[@billing_account.cc.length-4,4]
-         else
-           @billing_account.last4 = @billing_account.cc
-         end
-         @billing_account.encyrpt_cc
-       end
-
-       name_arr=@billing_account.name.split(" ")
-       fname=name_arr.shift
-       lname=name_arr.join(" ").strip
-
-       billing_address = {
-           :name => "#{@user.fname} #{@user.lname}",
-           :address1 => @billing_account.address1 ,
-           :zip =>@billing_account.zip,
-           :city     => @billing_account.city,
-           :state    => @billing_account.state,
-           :country  => 'US'
-       }
-
-       if @billing_account.cardtype=="telecheck"
-
-        linkpoint_options = {
-          :order_id => "",
-          :address => {}, 
-          :address1 => billing_address, 
-          :billing_address => billing_address, 
-          :ip=>"127.0.0.1", 
-          :telecheck_account => @billing_account.checking_account_number,
-          :telecheck_routing => @billing_account.bank_routing_number,
-          :telecheck_checknumber => params[:check_number], 
-          :telecheck_dl => @billing_account.drivers_license_number, 
-          :telecheck_dlstate => @billing_account.drivers_license_state, 
-          :telecheck_accounttype => @billing_account.checking_account_type
-        }
-
-          response = BILLING_GW.purchase(1, nil, linkpoint_options)
-          logger.info response.inspect
-
-          if response.params["approved"]=="SUBMITTED"
-            flash_message(:notice, "eCheck verified.")
-            @billing_account.save
-            account.update_attribute(:card_verified, true)
-            redirect_to :action=>"index"
-            return
-          else
-            flash_now(:error, "There was a problem validating your eCheck.  Please contact support for help. Error #{response.params["error"]}")
-          end
-
-       else
-         # test an auth to make sure this card is good.
-         creditcard = ActiveMerchant::Billing::CreditCard.new(
-           :number     => @billing_account.decrypt_cc,
-           :month      => @billing_account.expires_month,
-           :year       => @billing_account.expires_year,
-           :type       => @billing_account.cardtype,
-           :first_name => fname,
-           :last_name  => lname,
-           :verification_value => params[:code]
-         )
-
-         if !creditcard.valid?
-           if creditcard.expired?
-             flash_now(:error, "The card expiration date you entered was invalid. Please try again.")
-             @billing_account.cc = ""
-           else
-             flash_now(:error, "The card number or security code you entered was invalid. Please try again.")
-             @billing_account.cc = ""
-           end
-           return
-         end
-
-
-         options = {:address => {}, :address1 => billing_address, :billing_address => billing_address, :ip=>"127.0.0.1", :order_id=>""}
-         response = BILLING_GW.authorize(1, creditcard,options)
-         logger.info response.inspect
-
-         if response.success?
-           flash_message(:notice, "Card verified.")
-           @billing_account.save
-           account.update_attribute(:card_verified, true)
-           redirect_to :action=>"index"
-           return
-         else
-           flash_now(:error, "There was a problem validating your credit card.  Please email info@impactdialing.com for further support.")
-         end
-       end
-
-     end
-
-   end
-
- 
-
-  def campaign_clear_calls
-    ActiveRecord::Base.connection.execute("update voters set result=NULL, status='not called' where campaign_id=#{params[:id]}")
-    #    ActiveRecord::Base.connection.execute("delete from voter_results where campaign_id=#{params[:id]}")
-    flash_message(:notice, "Calls cleared")
-    redirect_to client_campaigns_path(params[:id])
-    return
-  end
-
-  def scripts
-    @breadcrumb="Scripts"
-    @scripts = @user.account.scripts.active.manual.paginate :page => params[:page], :order => 'name'
-  end
-
-
-  def voter_view
-    @campaign = Campaign.find_by_id_and_account_id(params[:campaign_id],account.id)
-    @breadcrumb=[{"Campaigns"=>client_campaigns_path},{@campaign.name => client_campaign_path(@campaign)},"View Voters"]
-    @campaign = account.campaigns.find_by_id(params[:campaign_id])
-    @breadcrumb=[{"Campaigns"=>"/client/campaigns"},{"#{@campaign.name}"=>client_campaign_path(@campaign)},"View Voters"]
-    @voters = Voter.paginate :page => params[:page], :conditions =>"active=1 and campaign_id=#{@campaign.id} and voter_list_id in (#{@campaign.voter_lists.collect{|c| c.id.to_s + ","}}0)", :order => 'LastName,FirstName,Phone'
-  end
-
-  def reports
-    if params[:id].blank?
-      @breadcrumb = "Reports"
-      @campaigns = account.campaigns.manual
-    else
-      @campaign = Campaign.find(params[:id])
-      @breadcrumb=[{"Reports"=>"/client/reports"},@campaign.name]
-    end
-  end
-
-
-  def report_realtime
-    check_warning
-    if params[:id].blank?
-      @breadcrumb = "Reports"
-    else
-      @campaign= account.campaigns.find_by_id(params[:id])
-      if @campaign.blank?
-        render :text=>"Unauthorized"
-        return
-      end
-      @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Realtime Report"]
-    end
-    #    require "#{Rails.root.to_s}/app/models/caller_session.rb"
-    #    require "#{Rails.root.to_s}/app/models/caller.rb"
-  end
-
-  def report_realtime_new
-    check_warning
-    if params[:timeframe].blank?
-      @timeframe = 10
-    else
-      @timeframe = params[:timeframe].to_i
-    end
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.blank?
-      render :text=>"Unauthorized"
-      return
-    end
-    @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"New Realtime Report"]
-  end
-
-  def update_report
-    #    Rails.logger.silence do
-    # CallerSession
-    # Caller
-    if params[:timeframe].blank?
-      @timeframe = 10
-    else
-      @timeframe = params[:timeframe].to_i
-    end
-
-    # if !params[:clear].blank?
-    #   cache_delete("avail_campaign_hash")
-    # end
-    # @avail_campaign_hash = cache_get("avail_campaign_hash") {{}}
-    @campaign = account.campaigns.find_by_id(params[:id])
-    render :layout=>false
-    #    end
-  end
-
-  def report_overview
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.blank?
-      render :text=>"Unauthorized"
-      return
-    end
-
-    set_report_date_range
-    @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Overview Report"]
-    sql = "#select distinct status from call_attempts
-
-    select
-    count(*) as cnt,
-    case WHEN ca.status='Call abandoned' THEN 'Call abandoned'
-      WHEN ca.status='Hangup or answering machine' THEN 'Hangup or answering machine'
-      WHEN ca.status='No answer' THEN 'No answer'
-      WHEN ca.status='No answer busy signal' THEN 'Busy signal'
-      ELSE ca.status
-      END AS result
-
-      from
-      call_attempts ca
-      where ca.campaign_id=#{@campaign.id}
-      and created_at > '#{@from_date.strftime("%Y-%m-%d")}'
-      and created_at < '#{(@to_date+1.day).strftime("%Y-%m-%d")}'
-      group by
-      case WHEN ca.status='Call abandoned' THEN 'Call abandoned'
-        WHEN ca.status='Hangup or answering machine' THEN 'Hangup or answering machine'
-        WHEN ca.status='No answer' THEN 'No answer'
-        WHEN ca.status='No answer busy signal' THEN 'Busy signal'
-        ELSE ca.status
-        END
-        order by count(*) desc"
-        @records = ActiveRecord::Base.connection.execute(sql)
-        @total=0
-        @records.each do |r|
-          @total = @total + r[0].to_i
-        end
-        @records.data_seek(0)
-
-        @voters_to_call = @campaign.voters_count("not called",false)
-        @voters_called = @campaign.voters_called
-        @totalvoters = @voters_to_call + @voters_called.length
-
-        @call_attempts = CallAttempt.find_all_by_campaign_id(@campaign.id)
-        @caller_sessions = CallerSession.find_all_by_campaign_id(@campaign.id)
-
-        @talkmins=0
-        @call_attempts.each do |attempt|
-          @talkmins += attempt.minutes_used
-        end
-        @callerMins=0
-        @caller_sessions.each do |session|
-          @callerMins += session.minutes_used
-        end
-
-  end
-
-  def report_overview_old
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.blank?
-      render :text=>"Unauthorized"
-      return
-    end
-    @script=@campaign.script
-    extra = ""
-
-    @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Answereds Call Report"]
-
-    set_report_date_range
-
-    sql = "
-        SELECT result_json
-        FROM call_attempts
-        where campaign_id=#{@campaign.id}
-        and result_json IS NOT NULL
-        and created_at > '#{@from_date.strftime("%Y-%m-%d")}'
-        and created_at < '#{(@to_date+1.day).strftime("%Y-%m-%d")}'  #{extra}
-        "
-        logger.info sql
-
-        @records = ActiveRecord::Base.connection.execute(sql)
-
-        @json_fields=[]
-        @results_hash={}
-        @names_hash={}
-        @records.each do |r|
-          this_record=YAML.load(r[0])
-          @json_fields = @json_fields | this_record.keys
-          this_record.keys.each do |f|
-            if this_record[f].index("name")
-            end
-#            @names_hash
-          end
-        end
-        # render :text=>json_fields.inspect
-        # return
-
-        @json_fields.each do |f|
-          @results_hash[f]={}
-          @records.data_seek(0)
-          @records.each do |r|
-            this_record=YAML.load(r[0])
-            if this_record.keys.index(f)
-              this_result = this_record[f]
-              # render :text=>this_record.inspect
-              # return
-#              this_result = this_records_fields[this_records_fields.index(f)]
-              @results_hash[f][this_result]=0 if !@results_hash[f].keys.index(this_result)
-              @results_hash[f][this_result] += 1
-            end
-          end
-        end
-
-#        render :text=>@results_hash.inspect
-#        return
-
-
-        @total=0
-        @records.each do |r|
-          @total = @total + r[0].to_i
-        end
-        @records.data_seek(0)
-
-        @voters_to_call = @campaign.voters_count("not called",false)
-        @voters_called = @campaign.voters_called
-        @totalvoters = @voters_to_call + @voters_called.length
-
-  end
-
-  # def show_memcached
-  #   @avail_campaign_hash = cache_get("avail_campaign_hash") {{}}
-  # end
-  def script_delete
-    @script = account.scripts.find_by_id(params[:id])
-    unless @script.blank?
-      campaign = account.campaigns.active.find_by_script_id(@script.id)
-      if campaign.nil?
-        @script.active=false
-        @script.save
-        flash_message(:notice, "Script deleted")
-      else
-        flash_message(:notice, I18n.t(:script_cannot_be_deleted))
-      end
-    end
-    redirect_to :back
-  end
-
-  def report_caller
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.blank?
-      render :text=>"Unauthorized"
-      return
-    end
-    if params[:type]=="1"
-      extra = "and result is not null"
-    end
-
-    @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Caller Report"]
-
-    set_report_date_range
-    caller_ids=CallerSession.all(:select=>"distinct caller_id", :conditions=>"campaign_id=#{@campaign.id}")
-    @callers=[]
-    caller_ids.each do |caller_session|
-      @callers<< Caller.find(caller_session.caller_id)
-    end
-
-    @responses = Voter.all(:select=>"distinct result", :conditions=>"campaign_id = #{@campaign.id} and result is not null and result_date > '#{@from_date.strftime("%Y-%m-%d")}' and result_date < '#{(@to_date+1.day).strftime("%Y-%m-%d")}'")
-    @num_responses = Voter.all(:conditions=>"campaign_id = #{@campaign.id} and result is not null and result_date > '#{@from_date.strftime("%Y-%m-%d")}' and result_date < '#{(@to_date+1.day).strftime("%Y-%m-%d")}'").length
-  end
-
-  def report_caller_overview
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.blank?
-      render :text=>"Unauthorized"
-      return
-    end
-    if params[:type]=="1"
-      extra = "and result is not null"
-    end
-
-    @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Caller Report"]
-
-    set_report_date_range
-    caller_ids=CallerSession.all(:select=>"distinct caller_id", :conditions=>"campaign_id=#{@campaign.id}")
-    @callers=[]
-    caller_ids.each do |caller_session|
-      @callers<< Caller.find(caller_session.caller_id)
-    end
-
-  end
-
-  def report_login
-
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.blank?
-      render :text=>"Unauthorized"
-      return
-    end
-    if params[:type]=="1"
-      extra = "and result is not null"
-    end
-
-    @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Caller Report"]
-    #      @logins = CallerSession.find_all_by_campagin_id(@campagin.id, :order=>"id desc")
-    @logins = CallerSession.find_all_by_campaign_id(@campaign.id, :order=>"id desc")
-  end
-
-  def report_download
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.blank?
-      render :text=>"Unauthorized"
-      return
-    end
-
-    @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Download Report"]
-    set_report_date_range
-
-    if params[:download]=="1"
-      #      attempts = CallAttempt.find_all_by_campaign_id(@campaign.id)
-
-      subsql = "select voter_id from call_attempts ca
-           join voters v on v.id=ca.voter_id
-           where ca.campaign_id=#{@campaign.id} and last_call_attempt_id=ca.id
-           and ca.created_at > '#{@from_date.strftime("%Y-%m-%d")}'
-           and ca.created_at < '#{(@to_date+1.day).strftime("%Y-%m-%d")}'
-      "
-      voters = ActiveRecord::Base.connection.execute(subsql)
-      voter_ids = []
-      voters.each do |a|
-        voter_ids << a[0]
-      end
-      if voter_ids.length==0
-        voter_subq = "0"
-      else
-        voter_subq = voter_ids.join(",")
-      end
-
-      sql = <<-EOSQL
-        select
-        ca.result, ca.result_digit , v.Phone, v.CustomID, v.LastName, v.FirstName, v.MiddleName, v.Suffix, v.Email, c.pin, c.name,  c.email, ca.status, ca.connecttime, ca.call_end, v.last_call_attempt_id=ca.id as final , ca.result_json, f.CustomID, f.LastName, f.FirstName, f.MiddleName, f.Suffix, f.Email, family_id_answered, cs.caller_number
-        from call_attempts ca
-        join voters v on v.id=ca.voter_id
-        left outer join callers c on c.id=ca.caller_id
-        left outer join caller_sessions cs on cs.id=ca.caller_session_id
-        left outer join families f on f.id=v.family_id_answered
-        where
-        ca.campaign_id=#{@campaign.id}
-        and v.id in (#{voter_subq})
-        order by v.id asc, ca.id asc
-      EOSQL
-      attempts = ActiveRecord::Base.connection.execute(sql)
-      logger.info "attempts: #{attempts}"
-      json_fields=[]
-      attempts.each do |a|
-        if json_fields.empty? && a[16]!=nil
-          json_fields = YAML.load(a[16]).keys
-        end
-      end
-      attempts.data_seek(0)
-
-      csv_string = CSV.generate do |csv|
-        #            csv << ["result", "result digit" , "voter phone", "voter id", "voter last", "voter first", "voter middle", "voter suffix", "voter email","caller pin", "caller name",  "caller email","status", "call start", "call end", "number attempts"]
-        #csv << ["id", "LastName", "FirstName", "MiddleName", "Suffix", "Phone", "Result", "Caller Name", "Status", "Call Start", "Call End", "Number Calls"] + json_fields #+ ["fam_id", "fam_LastName", "fam_FirstName", "fam_MiddleName", "fam_Suffix", "fam_Email"]
-        csv << ["id", "LastName", "FirstName", "MiddleName", "Suffix", "Phone", "Caller Name", "Caller Phone", "Status", "Call Start", "Call End", "Number Calls"] + json_fields #+ ["fam_id", "fam_LastName", "fam_FirstName", "fam_MiddleName", "fam_Suffix", "fam_Email"]
-        num_call_attempts=0
-        attempts.each do |a|
-          num_call_attempts+=1
-          #logger.info "a[15]: #{a[15]}"
-          if a[15]=="1"
-            #final attempt
-            #logger.info a.inspect
-            json_to_add=[]
-            if a[16].blank?
-              json_fields.each do |j|
-                json_to_add << ""
-              end
-            else
-              json=YAML.load(a[16])
-              json_fields.each do |j|
-                if json.keys.index(j)
-                  json_to_add << json[j]
-                else
-                  json_to_add << ""
-                end
-              end
-            end
-            #csv << [a[0],a[1],a[2],a[3],a[4],a[5],a[6],a[7],a[8],a[9],a[10],a[11],a[12],a[13],a[14],a[15],num_call_attempts]
-            #csv << [a[3],a[4],a[5],a[6],a[7],a[2],a[0],a[10],a[12],a[13],a[14],num_call_attempts]  + json_to_add + [a[17],a[18],a[19],a[20],a[21]]
-            if a[23]==0 || a[23]=="" || a[23]==nil || a[23]=="0"
-              #no fam
-              #logger.info "no fam"
-              csv << [a[3],a[4],a[5],a[6],a[7],a[2],a[10],a[24],a[12],a[13],a[14],num_call_attempts]  + json_to_add #+ [a[17],a[18],a[19],a[20],a[21]]
-            else
-              #fam
-              #logger.info "fam: #{a[23]}"
-              csv << [a[17],a[18],a[19],a[20],a[7],a[2],a[10],a[24],a[12],a[13],a[14],num_call_attempts]  + json_to_add
-            end
-
-            num_call_attempts=0
-          end
-        end
-      end
-      send_data csv_string, :type => "text/csv",  :filename=>"report.csv", :disposition => 'attachment'
-      return
-    end
-
-  end
-
-  def report_real
-    check_warning
-    if params[:id].blank?
-      @breadcrumb = "Reports"
-    else
-      @campaign = account.campaigns.find_by_id(params[:id])
-      if @campaign.blank?
-        render :text=>"Unauthorized"
-        return
-      end
-      @breadcrumb=[{"Reports"=>"/client/reports"},{"#{@campaign.name}"=>"/client/reports/#{@campaign.id}"},"Realtime Report"]
-    end
-  end
-
-  def update_report_real
-    if params[:timeframe].blank?
-      @timeframe = 10
-    else
-      @timeframe = params[:timeframe].to_i
-    end
-
-    # if !params[:clear].blank?
-    #   cache_delete("avail_campaign_hash")
-    # end
-    # @avail_campaign_hash = cache_get("avail_campaign_hash") {{}}
-    @campaign = account.campaigns.find_by_id(params[:id])
-    if @campaign.nil?
-      render :text=>"Campaign not found or access not permitted"
-      return
-    end
-    render :layout=>false
-  end
-  
 
   def policies
     render 'home/policies'
