@@ -3,6 +3,7 @@ require Rails.root.join("lib/twilio_lib")
 class Campaign < ActiveRecord::Base
   include Deletable
 
+  attr_accessible :type, :name, :caller_id, :script_id, :acceptable_abandon_rate, :time_zone, :start_time, :end_time, :recycle_rate, :answering_machine_detect, :voter_lists_attributes, :use_recordings, :recording_id
 
   has_many :caller_sessions
   has_many :voter_lists, :conditions => {:active => true}
@@ -14,11 +15,14 @@ class Campaign < ActiveRecord::Base
   has_many :answers
   has_many :note_responses
   has_many :call_responses
+  has_many :caller_groups
   belongs_to :script
-  belongs_to :voicemail_script, :class_name => 'Script', :foreign_key => 'voicemail_script_id'
+  belongs_to :voicemail_script, :class_name => 'Script'
   belongs_to :account
   belongs_to :recording
   has_many :downloaded_reports
+
+  accepts_nested_attributes_for :voter_lists
 
   delegate :questions_and_responses, :to => :script
 
@@ -33,19 +37,25 @@ class Campaign < ActiveRecord::Base
   }
   scope :for_caller, lambda { |caller| {:include => [:caller_sessions], :conditions => {"caller_sessions.caller_id" => caller.id}}}
 
-  before_create :create_uniq_pin
-
   validates :name, :presence => true
-  validates :caller_id, :presence => true, :unless => :new_campaign
-  validates :caller_id, :numericality => {:on => :update}, :length => {:on => :update, :minimum => 10, :maximum => 10}, :unless => Proc.new{|campaign| campaign.caller_id && campaign.caller_id.start_with?('+')}
+  validates :caller_id, :presence => true
+  validates :caller_id, :numericality => {}, :length => {:minimum => 10, :maximum => 10}, :unless => Proc.new{|campaign| campaign.caller_id && campaign.caller_id.start_with?('+')}
+  validates :script, :presence => true
+  validates :type, :presence => true, :inclusion => {:in => ['Preview', 'Progressive', 'Predictive', 'Robo']}
+  validates :acceptable_abandon_rate,
+            :numericality => {:greater_than_or_equal_to => 0.01, :less_than_or_equal_to => 0.10},
+            :allow_blank => true
+  validates :recycle_rate, :presence => true, :numericality => true
+  validates :time_zone, :presence => true, :inclusion => {:in => ActiveSupport::TimeZone.zones_map.map {|z| z.first}}
+  validates :start_time, :presence => true
+  validates :end_time, :presence => true
   validate :set_caller_id_error_msg
   validate :campaign_type_changed
   validate :script_changed_called
+  validate :no_caller_assigned_on_deletion
   cattr_reader :per_page
   @@per_page = 25
 
-  before_validation :set_untitled_name
-  before_save :set_untitled_name
   before_validation :sanitize_caller_id
 
   module Type
@@ -59,10 +69,18 @@ class Campaign < ActiveRecord::Base
     new_record?
   end
 
+  def no_caller_assigned_on_deletion
+    if active_change == [true, false] && !callers.empty?
+      errors.add(:base, 'There are currently callers assigned to this campaign. Please assign them to another campaign before deleting this one.')
+    end
+  end
+
+
+
   def set_caller_id_error_msg
       if errors[:caller_id].any?
-        errors.add(:base, 'Your Caller ID must be a 10-digit North American phone number or begin with "+" and the country code.')
         errors[:caller_id].clear
+        errors.add(:base, 'ID must be a 10-digit North American phone number or begin with "+" and the country code.')
       end
     end
 
@@ -71,7 +89,7 @@ class Campaign < ActiveRecord::Base
       errors.add(:base, 'You cannot change dialing modes while callers are logged in.')
     end
   end
-  
+
   def script_changed_called
     if script_id_changed? && call_attempts.count > 0
       errors.add(:base, I18n.t(:script_cannot_be_modified))
@@ -83,34 +101,12 @@ class Campaign < ActiveRecord::Base
     type == Type::PREVIEW || type == Type::PROGRESSIVE
   end
 
-  def set_untitled_name
-    self.name = "Untitled #{account.campaigns.count + 1}" if self.name.blank?
-  end
-
   def sanitize_caller_id
     self.caller_id = Voter.sanitize_phone(self.caller_id)
   end
 
-  def create_uniq_pin
-    pin = nil
-
-    loop do
-      pin = rand.to_s[2..6]
-      break unless Campaign.find_by_campaign_id(pin)
-    end
-    self.campaign_id = pin
-  end
-
   def set_answering_machine_detect
     self.answering_machine_detect = self.use_recordings = self.robo? && !self.voicemail_script.nil?
-  end
-
-  def disable_voter_list
-    voter_lists.each do |voter_list|
-      voter_list.enabled = false
-      voter_list.save
-      voter_list.voters.update_all(enabled: false)
-    end
   end
 
 
@@ -171,7 +167,7 @@ class Campaign < ActiveRecord::Base
     answer_count = Answer.select("possible_response_id").where("campaign_id = ?", self.id).within(from_date, to_date).group("possible_response_id").count
     total_answers = Answer.from("answers use index (index_answers_count_question)").where("campaign_id = ?",self.id).within(from_date, to_date).group("question_id").count
     questions = Question.where("id in (?)",question_ids.collect{|q| q.question_id})
-    questions.each do |question|        
+    questions.each do |question|
       result[question.text] = question.possible_responses.collect { |possible_response| possible_response.stats(answer_count, total_answers) }
       result[question.text] << {answer: "[No response]", number: 0, percentage:  0} unless question.possible_responses.find_by_value("[No response]").present?
     end
@@ -198,11 +194,8 @@ class Campaign < ActiveRecord::Base
   def abandoned_calls_time(from_date, to_date)
     call_attempts.between(from_date, to_date).with_status([CallAttempt::Status::ABANDONED]).sum('ceil(TIMESTAMPDIFF(SECOND ,connecttime,call_end)/60)').to_i
   end
-  
+
   def cost_per_minute
     0.09
   end
-  
-
-
 end
