@@ -2,7 +2,7 @@ RAILS_ROOT = File.expand_path('../..', __FILE__)
 require File.join(RAILS_ROOT, 'config/environment')
 DIALER_ROOT = ENV['DIALER_ROOT'] || File.expand_path('..', __FILE__)
 FileUtils.mkdir_p(File.join(DIALER_ROOT, 'log'), :verbose => true)
-ActiveRecord::Base.logger = Logger.new(File.open(File.join(DIALER_ROOT, 'log', "dialer_#{RAILS_ENV}.log"), 'a'))
+ActiveRecord::Base.logger = Logger.new(File.open(File.join(DIALER_ROOT, 'log', "dialer_#{ENV['RAILS_ENV']}.log"), 'a'))
 require 'em-http-request'
 require "em-synchrony"
 require "em-synchrony/em-http"
