@@ -1,6 +1,7 @@
 require "spec_helper"
 
 describe Script do
+  
   it "restoring makes it active" do
     script = Factory(:script, :active => false)
     script.restore
@@ -46,31 +47,31 @@ describe Script do
       account = Factory(:account)
       @script = Script.default_script(account)
     end
-    
-  
+
+
     it "should have name Demo Script" do
       @script.name.should == "Demo script"
     end
-  
+
     it "should have FirstName, lastName and Phone as voter fields" do
       @script.voter_fields.should eq('["FirstName","LastName","Phone"]')
     end
-  
+
     it "should have a default note" do
       @script.notes.length.should eq(1)
-      @script.notes.first.note.should eq("What's your favorite feature?")  
+      @script.notes.first.note.should eq("What's your favorite feature?")
     end
-    
+
     it "should add a default question" do
       @script.questions.length.should eq(1)
       @script.questions.first.text.should eq('How do you like the predictive dialer so far?')
     end
-    
+
     it "should add a default question with 4 possible responses" do
       @script.questions.length.should eq(1)
       @script.questions.first.possible_responses.length.should eq(4)
     end
-    
+
   end
 
   describe "questions and responses" do
@@ -81,6 +82,23 @@ describe Script do
       response_2 = Factory(:possible_response, :question => question)
       another_response = Factory(:possible_response)
       script.questions_and_responses.should == {question.text => [response_1.value, response_2.value]}
+    end
+  end
+
+  describe "deletion" do
+    it "should not delete a script that is being used by a campaign" do
+      script = Factory(:script)
+      campaign = Factory(:campaign, script: script)
+      script.active = false
+      script.save.should be_false
+      script.errors[:base].should == [I18n.t(:script_cannot_be_deleted)]
+    end
+
+    it "should delete a script that is not used by any campaign" do
+      script = Factory(:script)
+      campaign = Factory(:campaign)
+      script.active = false
+      script.save.should be_true
     end
   end
 end
