@@ -18,21 +18,6 @@ describe VoterListJob do
       UserMailer.stub(:new).and_return(mailer)
     end
 
-    describe "requirements" do
-
-      it "needs a list name" do
-        job = VoterListJob.new(@separator, @json_csv_column_headers, @csv_to_system_map, '', '', @campaign.id, @account.id, nil, nil,"")
-        mailer.should_receive(:voter_list_upload)
-        job.perform['errors'].first.should include "Name can't be blank"
-      end
-
-      it "should not save a list if the user already has a list with the same name" do
-        Factory(:voter_list, :account => @account, :campaign_id => @campaign.id, :name => "abcd")
-        job = VoterListJob.new(@separator, @json_csv_column_headers, @csv_to_system_map, '', 'abcd', @campaign.id, @account.id, nil, nil,"")
-        mailer.should_receive(:voter_list_upload)
-        job.perform['errors'].should include "Name for this list is already taken."
-      end
-    end
 
     describe "after import" do
       before :each do
@@ -41,7 +26,8 @@ describe VoterListJob do
 
       it "saves all the voters in the csv according to the mappings" do
         Voter.delete_all
-        job = VoterListJob.new(@separator, @json_csv_column_headers, @csv_to_system_map, @csv_filename, 'abcd', @campaign.id, @account.id, nil, nil,"")
+        voter_list = Factory(:voter_list, separator: ",", headers: "[]", csv_to_system_map: {"Phone" => "Phone", "LAST" =>"LastName"}.to_json, s3path: @csv_filename, campaign_id: @campaign.id, account_id: @account.id)
+        job = VoterListJob.new(voter_list.id, nil, nil,"")
         mailer.should_receive(:voter_list_upload)
         s3.should_receive(:value).and_return(File.open("#{fixture_path}/files/valid_voters_list.csv").read)
 
@@ -51,13 +37,6 @@ describe VoterListJob do
         Voter.first.LastName.should == "Bar"
       end
 
-      it "removes the temporary file from disk" do
-        temp_filename = "#{Rails.root}/tmp/#{@csv_filename}"
-        job = VoterListJob.new(@separator, @json_csv_column_headers, @csv_to_system_map, @csv_filename, 'abcd', @campaign.id, @account.id, nil, nil,"")
-        mailer.should_receive(:voter_list_upload)
-        s3.should_receive(:value).and_return(File.open("#{fixture_path}/files/valid_voters_list.csv").read)
-        job.perform
-      end
     end
 
     describe "custom fields" do
@@ -65,7 +44,8 @@ describe VoterListJob do
         @csv_filename = "valid_voters_list_#{Time.now.to_i}_#{rand(999)}"
         custom_field = "Custom"
         Voter.delete_all
-        job = VoterListJob.new(@separator, ["Phone", "Custom"].to_json, {"Phone"=>"Phone", custom_field=>custom_field}, @csv_filename, 'abcd', @campaign.id, @account.id, nil, nil,"")
+        voter_list = Factory(:voter_list, separator: ",", headers: "[]", csv_to_system_map: {"Phone" => "Phone", custom_field=>custom_field}.to_json, s3path: @csv_filename, campaign_id: @campaign.id, account_id: @account.id)
+        job = VoterListJob.new(voter_list.id, nil, nil,"")
         mailer.should_receive(:voter_list_upload)
         s3.should_receive(:value).and_return(File.open("#{fixture_path}/files/voters_custom_fields_list.csv").read)
 
@@ -89,7 +69,9 @@ describe VoterListJob do
       end
 
       it "should flash an error" do
-        job = VoterListJob.new(@separator, @json_csv_column_headers, @csv_to_system_map, @csv_filename, 'bui', @campaign.id, @account.id, nil, nil,"")
+        voter_list = Factory(:voter_list, separator: ",", headers: "[]", csv_to_system_map: {"Phone" => "Phone"}.to_json, s3path: @csv_filename, campaign_id: @campaign.id, account_id: @account.id)
+        job = VoterListJob.new(voter_list.id, nil, nil,"")
+
         mailer.should_receive(:voter_list_upload)
         s3.should_receive(:value).and_return(File.open("#{fixture_path}/files/invalid_voters_list.csv").read)
 
@@ -97,7 +79,9 @@ describe VoterListJob do
       end
 
       it "should not save the voters list entry" do
-        job = VoterListJob.new(@separator,@json_csv_column_headers,@csv_to_system_map,@csv_filename,'hui',@campaign.id,@account.id,nil,nil,"")
+        voter_list = Factory(:voter_list, separator: ",", headers: "[]", csv_to_system_map: {"Phone" => "Phone"}.to_json, s3path: @csv_filename, campaign_id: @campaign.id, account_id: @account.id)
+        job = VoterListJob.new(voter_list.id, nil, nil,"")
+        
         mailer.should_receive(:voter_list_upload)
         s3.should_receive(:value).and_return(File.open("#{fixture_path}/files/invalid_voters_list.csv").read)
         job.perform
