@@ -21,6 +21,7 @@ Spork.prefork do
   Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
   Dir[Rails.root.join("spec/shared/**/*.rb")].each {|f| require f}
   #Dir[Rails.root.join("simulator/new_simulator.rb")].each {|f| require f}
+  
 
   RSpec.configure do |config|
     # == Mock Framework
@@ -31,7 +32,19 @@ Spork.prefork do
     # config.mock_with :flexmock
     # config.mock_with :rr
     config.mock_with :rspec
-
+    
+    config.before(:suite) do
+       DatabaseCleaner.strategy = :truncation
+    end
+    config.before(:each) do
+      DatabaseCleaner.start
+    end
+    
+    
+    config.after(:each) do
+        DatabaseCleaner.clean
+    end
+      
     # If you're not using ActiveRecord, or you'd prefer not to run each of your
     # examples within a transaction, remove the following line or assign false
     # instead of true.
